+++ src/cu_std/win32/tthread.h
@@ -15,7 +15,7 @@ struct TThreadContext{
 
 typedef HANDLE TSemaphore;
 
-TThreadContext TCreateThread(s32 (*call_fptr)(void*),u32 stack_size,void* args);
+TThreadContext TCreateThread(s64 (*call_fptr)(void*),u32 stack_size,void* args);
 
 TSemaphore TCreateSemaphore(u32 value = 0);
